integrate ngx-electron
