+++ src/app/app.component.ts
@@ -13,3 +13,4 @@ export class AppComponent {
     this.viewContainerRef = viewContainerRef;
   }
 }
+
