+++ src/app/shared/shared.module.ts
@@ -5,6 +5,7 @@ import { ControlMessagesComponent } from './control-messages/control-messages.co
 import { ValidationService } from './validation.service';
 import { ModelService } from './model.service';
 
+
 @NgModule({
     imports: [
       CommonModule,
@@ -19,7 +20,7 @@ import { ModelService } from './model.service';
   ],
   providers: [
     ValidationService,
-    ModelService
+    ModelService,
 
   ]
 })
